+++ include/micros_rtt/connection_base.hpp
@@ -16,23 +16,32 @@ public:
   
 private:
   std::string topic_;
-  bool is_interprocess_;
   ChannelElementBase::shared_ptr channel_element;
+  ChannelElementBase::shared_ptr mq_channel_element;
 
 public:
   ConnectionBase() {}
-  ConnectionBase(std::string topic, bool is_interprocess = false) : topic_(topic), is_interprocess_(is_interprocess) {}
+  ConnectionBase(std::string topic) : topic_(topic) {}
   ~ConnectionBase() {}
   
   std::string getTopic() {return topic_;}
   
   ChannelElementBase::shared_ptr getChannelElement() {return channel_element;}
+  ChannelElementBase::shared_ptr getMQChannelElement() {return mq_channel_element;}
   
-  bool addConnection(ChannelElementBase::shared_ptr channel) {channel_element = channel;}
-  
-  bool isInterprocess() {return is_interprocess_;}
+  bool addConnection(ChannelElementBase::shared_ptr channel) 
+  {
+    channel_element = channel;
+    return true;
+  }
+  bool addMQConnection(ChannelElementBase::shared_ptr channel) 
+  {
+    mq_channel_element = channel;
+    return true;
+  }
 
-  virtual bool channelReady(ChannelElementBase::shared_ptr channel){}
+  virtual bool channelReady(ChannelElementBase::shared_ptr channel) = 0;
+  virtual bool mqChannelReady(ChannelElementBase::shared_ptr channel) = 0;
 };
 
 }