+++ include/micros_rtt/subscriber.h
@@ -8,8 +8,9 @@
 #ifndef MICROSRTT_SUBSCRIBER_HANDLE_H
 #define MICROSRTT_SUBSCRIBER_HANDLE_H
 
-#include "ros/subscriber.h"
+#include "ros/ros.h"
 #include "topic_manager.h"
+#include "subscription.h"
 
 namespace micros_rtt
 {
@@ -17,27 +18,21 @@ namespace micros_rtt
 class Subscriber 
 {
 public:
-  Subscriber();
-  Subscriber(ConnectionBasePtr sub_connection, ros::Subscriber ros_subscriber);
-  ~Subscriber();
+  Subscriber(){}
+  Subscriber(ros::Subscriber ros_subscriber, ConnectionBasePtr sub_connection);
+  ~Subscriber(){}
   
   template<class M> void call() 
   { 
     if(subscription)
     {
-	  ROS_INFO("have subscription %s", subscription->getTopic().c_str());
-	  if (subscription->isInterprocess())
-	  {
-        InterSubscription<M>* sub = static_cast< InterSubscription<M>* >(subscription.get ());
-        sub->call();
-	  }
-	  else
-	  {
-	    Subscription<M>* sub = static_cast< Subscription<M>*>(subscription.get());
-	    sub->call();
-	  }		  
+	    ROS_DEBUG("micros subscriber have subscription %s", subscription->getTopic().c_str());
+  	  Subscription<M>* sub = static_cast< Subscription<M>*>(subscription.get());
+  	  sub->call();
     }
   }
+  
+  ros::Subscriber getRosSubscriber() { return ros_sub; }
   
 private:
   ConnectionBasePtr subscription;