+++ include/micros_rtt/subscription.h
@@ -1,9 +1,10 @@
 #ifndef MICROSRTT_SUBSCRIPTION_H
 #define MICROSRTT_SUBSCRIPTION_H
 
-#include "oro/channel_data_element.hpp"
+#include "ros/ros.h"
 #include "boost/function.hpp"
-#include "connection_base.hpp"
+#include "micros_rtt/connection_base.hpp"
+#include "micros_rtt/oro/channel_data_element.hpp"
 
 namespace micros_rtt
 {  
@@ -25,10 +26,21 @@ public:
     callback = fp;
   }
   
-  bool channelReady( ChannelElementBase::shared_ptr end_port) 
+  bool channelReady( ChannelElementBase::shared_ptr channel) 
   {
-    if (end_port->inputReady ()) 
+    if (channel && channel->inputReady())
     {
+      addConnection(channel);
+      return true;
+    }
+    return false;
+  }
+  
+  bool mqChannelReady( ChannelElementBase::shared_ptr channel) 
+  {
+    if (channel && channel->inputReady())
+    {
+      addMQConnection(channel);
       return true;
     }
     return false;
@@ -38,22 +50,43 @@ public:
   {
     FlowStatus result;
     M sample;
+    M mq_sample;
     typename ChannelElement<M>::shared_ptr input = static_cast< ChannelElement<M>* >( this->getChannelElement().get() );
-    if ( input ) 
+    typename ChannelElement<M>::shared_ptr mq_input = static_cast< ChannelElement<M>* >( this->getMQChannelElement().get() );
+    
+    if (!(mq_input || input))
     {
-      FlowStatus tresult = input->read(sample, false);
+      return false;
+    }
+    //if ( input ) 
+    //{
+    //  FlowStatus tresult = input->read(sample, false);
+    //  // the result trickery is for not overwriting OldData with NoData.
+    //  if (tresult == NewData) 
+    //  {
+    //    result = tresult;
+    //    callback(sample);
+    //  }
+    //  // stores OldData result
+    //  if (tresult > result)
+    //    result = tresult;
+    //}
+    
+    if ( mq_input ) 
+    {
+      FlowStatus tresult = mq_input->read(mq_sample, false);
       // the result trickery is for not overwriting OldData with NoData.
       if (tresult == NewData) 
       {
         result = tresult;
-        callback(sample);
-        return true;
+        //callback(sample);
       }
       // stores OldData result
       if (tresult > result)
         result = tresult;
     }
-    return false;
+    return result;
+
   }
 private:
   boost::function<void(M)> callback;