+++ src/subscriber.cpp
@@ -10,18 +10,10 @@
 namespace micros_rtt
 {
  
-Subscriber::Subscriber() 
-{ 
-}
- 
-Subscriber::Subscriber(ConnectionBasePtr sub_connection, ros::Subscriber ros_subscriber) 
+Subscriber::Subscriber(ros::Subscriber ros_subscriber, ConnectionBasePtr sub_connection) 
 {
   subscription = sub_connection;
   ros_sub = ros_subscriber;
-}
-
-Subscriber::~Subscriber() 
-{ 
 }
 
 }