+++ CHANGELOG.rst
@@ -2,6 +2,13 @@
 Changelog for package micros_rtt
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+0.0.5 (2015-09-21)
+------------------
+* Add inter-process transport for local comunication.
+* optimize code structure.
+* fix exist bugs.
+* Contributors: sukha, sukha-cn
+
 0.0.4 (2015-07-22)
 ------------------
 * change the prefix of package name from ros to hpcl